# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ClaimJobRequest": {
        "additionalProperties": false,
        "example": {
          "max_types": [
            "create_pond",
            "backup_pond"
          ]
        },
        "properties": {
          "max_types": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/JobType"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Types"
          }
        },
        "title": "ClaimJobRequest",
        "type": "object"
      },
      "ClaimJobResponse": {
        "additionalProperties": false,
        "example": {
          "job": {
            "id": "77777777-7777-7777-7777-777777777777",
            "payload": {
              "cpus": 0.5,
              "db_password_plain": "p0nd-Temp!2026",
              "host_port": 15007,
              "image": "postgres:16-alpine",
              "memory_mb": 512,
              "name": "inventario-demo"
            },
            "pond_id": "66666666-6666-6666-6666-666666666666",
            "type": "create_pond"
          }
        },
        "properties": {
          "job": {
            "$ref": "#/components/schemas/ClaimedJob"
          }
        },
        "required": [
          "job"
        ],
        "title": "ClaimJobResponse",
        "type": "object"
      },
      "ClaimedJob": {
        "additionalProperties": false,
        "example": {
          "id": "77777777-7777-7777-7777-777777777777",
          "payload": {
            "cpus": 0.5,
            "db_password_plain": "p0nd-Temp!2026",
            "host_port": 15007,
            "image": "postgres:16-alpine",
            "memory_mb": 512,
            "name": "inventario-demo"
          },
          "pond_id": "66666666-6666-6666-6666-666666666666",
          "type": "create_pond"
        },
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/ClaimedJobPayload"
          },
          "pond_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pond Id"
          },
          "type": {
            "$ref": "#/components/schemas/JobType"
          }
        },
        "required": [
          "id",
          "type",
          "payload"
        ],
        "title": "ClaimedJob",
        "type": "object"
      },
      "ClaimedJobPayload": {
        "additionalProperties": false,
        "example": {
          "cpus": 0.5,
          "db_password_plain": "p0nd-Temp!2026",
          "host_port": 15007,
          "image": "postgres:16-alpine",
          "memory_mb": 512,
          "name": "inventario-demo"
        },
        "properties": {
          "cpus": {
            "title": "Cpus",
            "type": "number"
          },
          "db_password_plain": {
            "title": "Db Password Plain",
            "type": "string"
          },
          "host_port": {
            "title": "Host Port",
            "type": "integer"
          },
          "image": {
            "title": "Image",
            "type": "string"
          },
          "memory_mb": {
            "title": "Memory Mb",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "host_port",
          "memory_mb",
          "cpus",
          "db_password_plain",
          "image"
        ],
        "title": "ClaimedJobPayload",
        "type": "object"
      },
      "CompleteJobRequest": {
        "additionalProperties": false,
        "example": {
          "result": {
            "observed_state": "running"
          },
          "status": "succeeded"
        },
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "enum": [
              "succeeded",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "CompleteJobRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HeartbeatContainer": {
        "additionalProperties": false,
        "example": {
          "name": "pond-inventario-demo",
          "state": "running",
          "uptime": "2h"
        },
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "uptime": {
            "title": "Uptime",
            "type": "string"
          }
        },
        "required": [
          "name",
          "state",
          "uptime"
        ],
        "title": "HeartbeatContainer",
        "type": "object"
      },
      "HeartbeatRequest": {
        "additionalProperties": false,
        "example": {
          "containers": [
            {
              "name": "pond-inventario-demo",
              "state": "running",
              "uptime": "2h"
            }
          ],
          "samples": [
            {
              "container_state": "running",
              "pond_name": "inventario-demo",
              "size_bytes": 124901
            }
          ]
        },
        "properties": {
          "containers": {
            "items": {
              "$ref": "#/components/schemas/HeartbeatContainer"
            },
            "title": "Containers",
            "type": "array"
          },
          "samples": {
            "items": {
              "$ref": "#/components/schemas/HeartbeatSample"
            },
            "title": "Samples",
            "type": "array"
          }
        },
        "required": [
          "containers",
          "samples"
        ],
        "title": "HeartbeatRequest",
        "type": "object"
      },
      "HeartbeatSample": {
        "additionalProperties": false,
        "example": {
          "container_state": "running",
          "pond_name": "inventario-demo",
          "size_bytes": 124901
        },
        "properties": {
          "container_state": {
            "title": "Container State",
            "type": "string"
          },
          "pond_name": {
            "title": "Pond Name",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "pond_name",
          "size_bytes",
          "container_state"
        ],
        "title": "HeartbeatSample",
        "type": "object"
      },
...
      "JobType": {
        "enum": [
          "create_pond",
          "start_pond",
          "stop_pond",
          "delete_pond",
          "backup_pond",
          "restore_pond"
        ],
        "title": "JobType",
        "type": "string"
      },
      "JsonRpcRequest": {
        "additionalProperties": false,
        "example": {
          "id": "tool-1",
          "jsonrpc": "2.0",
          "method": "tools/call",
          "params": {
            "arguments": {},
            "name": "list_ponds"
          }
        },
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "jsonrpc": {
            "const": "2.0",
            "title": "Jsonrpc",
            "type": "string"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          }
        },
        "required": [
          "jsonrpc",
          "method"
        ],
        "title": "JsonRpcRequest",
        "type": "object"
      },
      "JsonRpcResponse": {
        "additionalProperties": false,
        "example": {
          "id": "tool-1",
          "jsonrpc": "2.0",
          "result": {
            "message": "MCP skeleton ready",
            "status": "ok"
          }
        },
        "properties": {
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "jsonrpc": {
            "const": "2.0",
            "title": "Jsonrpc",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          }
        },
        "required": [
          "jsonrpc"
        ],
        "title": "JsonRpcResponse",
        "type": "object"
      },
...
      "MCPInfoResponse": {
        "additionalProperties": false,
        "example": {
          "enabled": true,
          "endpoint": "/mcp",
          "mutable_tools": [
            "create_pond",
            "delete_pond",
            "restore_backup",
            "run_sql",
            "confirm_action",
            "cancel_confirmation"
          ],
          "prompt_summary": "Las tools mutantes requieren explicaci\u00f3n previa y confirmaci\u00f3n expl\u00edcita.",
          "read_only_tools": [
            "whoami",
            "list_ponds",
            "get_pond",
            "get_connection",
            "list_subscriptions",
            "list_backups",
            "get_usage"
          ]
        },
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "endpoint": {
            "title": "Endpoint",
            "type": "string"
          },
          "mutable_tools": {
            "items": {
              "type": "string"
            },
            "title": "Mutable Tools",
            "type": "array"
          },
          "prompt_summary": {
            "title": "Prompt Summary",
            "type": "string"
          },
          "read_only_tools": {
            "items": {
              "type": "string"
            },
            "title": "Read Only Tools",
            "type": "array"
          }
        },
        "required": [
          "enabled",
          "endpoint",
          "read_only_tools",
          "mutable_tools",
          "prompt_summary"
        ],
        "title": "MCPInfoResponse",
        "type": "object"
      },
...
      "NodeSampleUpload": {
        "additionalProperties": false,
        "example": {
          "container_state": "running",
          "pond_id": "66666666-6666-6666-6666-666666666666",
          "size_bytes": 124901
        },
        "properties": {
          "container_state": {
            "title": "Container State",
            "type": "string"
          },
          "pond_id": {
            "format": "uuid",
            "title": "Pond Id",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "pond_id",
          "size_bytes",
          "container_state"
        ],
        "title": "NodeSampleUpload",
        "type": "object"
      },
      "OkResponse": {
        "additionalProperties": false,
        "example": {
          "ok": true
        },
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "title": "OkResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "KoiCloud Phase 0 backend contract freeze",
    "title": "KoiCloud API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/v1/heartbeat": {
      "post": {
        "operationId": "internal_heartbeat",
        "parameters": [
          {
            "in": "header",
            "name": "X-Node-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Node-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Heartbeat",
        "tags": [
          "internal"
        ]
      }
    },
    "/internal/v1/jobs/claim": {
      "post": {
        "operationId": "internal_claim_job",
        "parameters": [
          {
            "in": "header",
            "name": "X-Node-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Node-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "204": {
            "description": "No hay jobs disponibles"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Claim Job",
        "tags": [
          "internal"
        ]
      }
    },
    "/internal/v1/jobs/{job_id}/complete": {
      "post": {
        "operationId": "internal_complete_job",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Node-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Node-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Job",
        "tags": [
          "internal"
        ]
      }
    },
    "/internal/v1/samples": {
      "post": {
        "operationId": "internal_ingest_samples",
        "parameters": [
          {
            "in": "header",
            "name": "X-Node-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Node-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "examples": [
                  {
                    "container_state": "running",
                    "pond_id": "66666666-6666-6666-6666-666666666666",
                    "size_bytes": 124901
                  }
                ],
                "items": {
                  "$ref": "#/components/schemas/NodeSampleUpload"
                },
                "title": "Payload",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Samples",
        "tags": [
          "internal"
        ]
      }
    },
    "/mcp": {
      "get": {
        "operationId": "mcp_info",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-KOI-Agent-Password",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Koi-Agent-Password"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPInfoResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mcp Info",
        "tags": [
          "mcp"
        ]
      },
      "post": {
        "operationId": "mcp_jsonrpc",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-KOI-Agent-Password",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Koi-Agent-Password"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonRpcRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonRpcResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mcp Jsonrpc",
        "tags": [
          "mcp"
        ]
      }
    }
  }
}